+++ retention/disable_exposures_ttl.sql
@@ -1,0 +1,2 @@
+ALTER TABLE {{exposures_table}}
+REMOVE TTL