+++ retention/enable_conversions_ttl.sql
@@ -1,0 +1,2 @@
+ALTER TABLE {{conversions_table}}
+MODIFY TTL ts + INTERVAL {{retention_days}} DAY