+++ retention/enable_exposures_ttl.sql
@@ -1,0 +1,2 @@
+ALTER TABLE {{exposures_table}}
+MODIFY TTL ts + INTERVAL {{retention_days}} DAY